requirements.txt
faster-whisper==1.2.1
psutil==5.9.8
fastapi
uvicorn
numpy
sounddevice
scipy
pywebview

